+++ src/main/java/ru/kata/spring/boot_security/demo/services/RoleServiceImpl.java
@@ -23,13 +23,13 @@ public class RoleServiceImpl implements RoleService {
     public List<Role> getSetOfRoles(List<String> rolesId) {
         List<Role> roles = new ArrayList<>();
         for (String id : rolesId) {
-            roles.add(getRoleById(Integer.parseInt(id)));
+            roles.add(getRoleById(Long.parseLong(id)));
         }
         return roles;
     }
 
     @Override
-    public Role getRoleById(int id) {
+    public Role getRoleById(long id) {
         return roleRepository.findById(id).orElse(null);
     }
